+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 2.40.0 (2023-02-23)
 - Added support for PAT problems in the NowCoder parser
 - Added support for the vjudge.csgrandeur.cn mirror to the Virtual Judge parsers